Added discovery strategy.
The robot now remembers what connections
have been discovered. When choosing a direction
to drive in, unvisited paths are visited.
If there are no unknown paths at the current
intersection, the robot drives to the closest
node that has unvisited paths using pathfinding
with dijkstras algorithm.

When no undiscovered paths are left, the robot
starts to drive randomly.